% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-projects.R
\name{find_project_root}
\alias{find_project_root}
\title{Find the root of a Quarto project}
\usage{
find_project_root(path = ".")
}
\arguments{
\item{path}{Character. Path to check for Quarto project files. Defaults to
current working directory.}
}
\value{
Character Path of the project root directory if found, or \code{NULL}
}
\description{
This function checks if the current working directory is within a Quarto
project by looking for Quarto project files (\verb{_quarto.yml} or \verb{_quarto.yaml}).
Unlike \code{\link[=is_running_quarto_project]{is_running_quarto_project()}}, this works both during rendering and
interactive sessions.
}
\examples{
\dontshow{if (quarto_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dir <- tempfile()
dir.create(dir)
find_project_root(dir)
quarto_create_project(dir)
find_project_root(dir)

xfun::in_dir(dir,
  # Check if current directory is in a Quarto project
  !is.null(find_project_root())
)
# clean up
unlink(dir, recursive = TRUE)

\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=is_running_quarto_project]{is_running_quarto_project()}} for detecting active Quarto rendering
}
