% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{find_quarto}
\alias{find_quarto}
\title{Locate the quarto executable or abort}
\usage{
find_quarto()
}
\value{
A string with the path to the quarto executable. If not found this
function aborts with a user-friendly error message.
}
\description{
A small helper that wraps \code{\link[=quarto_path]{quarto_path()}} and throws a helpful error if the
quarto CLI cannot be found. This is used internally by other functions that
need to call the Quarto binary.
}
\examples{
if (interactive()) {
  # returns path or aborts
  try(find_quarto(), silent = TRUE)
}

}
\keyword{internal}
