% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-projects.R
\name{is_quarto_project}
\alias{is_quarto_project}
\title{Check if working within a Quarto project structure}
\usage{
is_quarto_project(path = ".")
}
\arguments{
\item{path}{Character. Path to check for Quarto project files. Defaults to
current working directory.}
}
\value{
Logical indicating if a Quarto project structure is detected
}
\description{
This function checks if the current working directory is within a Quarto
project by looking for Quarto project files (\verb{_quarto.yml} or \verb{_quarto.yaml}).
Unlike \code{\link[=is_running_quarto_project]{is_running_quarto_project()}}, this works both during rendering and
interactive sessions.
}
\examples{
\dontshow{if (quarto_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dir <- tempfile()
dir.create(dir)
is_quarto_project(dir)
quarto_create_project(dir)
is_quarto_project(dir)

xfun::in_dir(dir,
  # Check if current directory is in a Quarto project
  is_quarto_project()
)
# clean up
unlink(dir, recursive = TRUE)

\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=is_running_quarto_project]{is_running_quarto_project()}} for detecting active Quarto rendering
}
