% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{is_using_quarto}
\alias{is_using_quarto}
\title{Check is a directory is using quarto}
\usage{
is_using_quarto(dir = ".", verbose = FALSE)
}
\arguments{
\item{dir}{The directory to check}

\item{verbose}{print message about the result of the check}
}
\description{
This function will check if a directory is using quarto by looking for
\itemize{
\item \verb{_quarto.yml} at its root
\item at least one \code{.qmd} file in the directory
}
}
\examples{
dir.create(tmpdir <- tempfile())
is_using_quarto(tmpdir)
file.create(file.path(tmpdir, "_quarto.yml"))
is_using_quarto(tmpdir)
unlink(tmpdir, recursive = TRUE)
}
