% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog.R
\name{new_blog_post}
\alias{new_blog_post}
\title{Create a new blog post}
\usage{
new_blog_post(
  title,
  dest = NULL,
  open = rlang::is_interactive(),
  call = rlang::current_env(),
  ...
)
}
\arguments{
\item{title}{A character string for the title of the post. It is converted
to title case via \code{\link[tools:toTitleCase]{tools::toTitleCase()}}.}

\item{dest}{A character string (or NULL) for the path within \code{posts}. By
default, the title is adapted as the directory name.}

\item{open}{A logical: have the default editor open a window to edit the
\code{index.qmd} file?}

\item{call}{A call object for reporting errors.}

\item{...}{A named list of values to be added to the yaml header, such as
\code{description}, \code{author}, \code{categories}, etc.}
}
\value{
The path to the index file.
}
\description{
Creates (and potentially opens) the \code{index.qmd} file for a new blog post.
}
\examples{
\dontrun{
 \donttest{
new_blog_post("making quarto blog posts", categories = c("R"))

 }
}

}
