% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-package.R
\docType{package}
\name{quarto-package}
\alias{quarto}
\alias{quarto-package}
\title{quarto: R Interface to 'Quarto' Markdown Publishing System}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convert R Markdown documents and 'Jupyter' notebooks to a variety of output formats using 'Quarto'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quarto-dev/quarto-r}
  \item \url{https://quarto-dev.github.io/quarto-r/}
  \item Report bugs at \url{https://github.com/quarto-dev/quarto-r/issues}
}

}
\author{
\strong{Maintainer}: Christophe Dervieux \email{cderv@posit.co} (\href{https://orcid.org/0000-0003-4474-2498}{ORCID})

Authors:
\itemize{
  \item Christophe Dervieux \email{cderv@posit.co} (\href{https://orcid.org/0000-0003-4474-2498}{ORCID})
  \item JJ Allaire \email{jj@posit.co} (\href{https://orcid.org/0000-0003-0174-9868}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Gordon Woodhull [contributor]
}

}
\keyword{internal}
