% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-package.R
\docType{data}
\name{quarto}
\alias{quarto}
\title{Internal package state}
\format{
An object of class \code{environment} of length 0.
}
\usage{
quarto
}
\description{
Internal package state
}
\keyword{datasets}
