% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{quarto_binary_sitrep}
\alias{quarto_binary_sitrep}
\title{Check configurations for quarto binary used}
\usage{
quarto_binary_sitrep(verbose = TRUE, debug = FALSE)
}
\arguments{
\item{verbose}{if \code{FALSE}, only return the result of the check.}

\item{debug}{if \code{TRUE}, print more information about value set in configurations.}
}
\value{
\code{TRUE} if this package should be using the same quarto binary as the
RStudio IDE. \code{FALSE} otherwise if a difference is detected or quarto is not
found. Use \code{verbose = TRUE} or\code{debug = TRUE} to get detailed information.
}
\description{
This function check the configuration for the quarto package R package to
detect a possible difference with version used by RStudio IDE.
}
\examples{
quarto_binary_sitrep(verbose = FALSE)
quarto_binary_sitrep(verbose = TRUE)
quarto_binary_sitrep(debug = TRUE)

}
