% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{quarto_install_extension}
\alias{quarto_install_extension}
\title{Quarto Install Extension}
\usage{
quarto_install_extension(template = NULL)
}
\arguments{
\item{extension}{The extension to install, either an archive or a GitHub
repository as described in the documentation
\url{https://quarto.org/docs/extensions/formats.html}.}
}
\description{
Install an extension into a Quarto project.
}
\examples{
\dontrun{
# Install a template and set up a draft document from a GitHub repository
quarto_install_extension("quarto-ext/fontawesome")

# Install a template and set up a draft document from a ZIP archive
quarto_install_extension("https://github.com/quarto-ext/fontawesome/archive/refs/heads/main.zip")
}

}
