% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{quarto_list_extensions}
\alias{quarto_list_extensions}
\title{List Installed Quarto extensions}
\usage{
quarto_list_extensions(quiet = FALSE, quarto_args = NULL)
}
\arguments{
\item{quiet}{Suppress warning and other messages.}

\item{quarto_args}{Character vector of other \code{quarto} CLI arguments to append
to the Quarto command executed by this function. This is mainly intended for
advanced usage and useful for CLI arguments which are not yet mirrored in a
dedicated parameter of this \R function. See \verb{quarto render --help} for options.}
}
\description{
List Quarto Extensions in this folder or project by running \verb{quarto list}
}
\examples{
\dontrun{
# List Quarto Extensions in this folder or project
quarto_list_extensions()
}

}
