% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-logging.R
\name{quarto_log_level}
\alias{quarto_log_level}
\title{Return the current log level for Quarto
unless level is provided, in which case
it checks if the current log level matches the provided level.}
\usage{
quarto_log_level(level = NULL)
}
\description{
Return the current log level for Quarto
unless level is provided, in which case
it checks if the current log level matches the provided level.
}
