% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{quarto_metadata}
\alias{quarto_metadata}
\title{Read Quarto Metadata}
\usage{
quarto_metadata(input)
}
\arguments{
\item{input}{The input file or project directory to read metadata for.}
}
\value{
Named list with metadata. For input files, the named list
is keyed by output format. For projects, all project level metadata
defined in _quarto is contained in the list.
}
\description{
Read YAML metadata for an input file or project.
}
\examples{
\dontrun{
# Read metadata for file
quarto_metadata("notebook.Rmd")

# Read metadata for project
quarto_metdata("myproject")
}

}
