% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{quarto_remove_extension}
\alias{quarto_remove_extension}
\title{Remove a Quarto extensions}
\usage{
quarto_remove_extension(
  extension = NULL,
  no_prompt = FALSE,
  quiet = FALSE,
  quarto_args = NULL
)
}
\arguments{
\item{extension}{The extension to remove, either an archive or a GitHub
repository as described in the documentation
\url{https://quarto.org/docs/extensions/managing.html}.}

\item{no_prompt}{Do not prompt to confirm approval to download external extension.}

\item{quiet}{Suppress warning and other messages.}

\item{quarto_args}{Character vector of other \code{quarto} CLI arguments to append
to the Quarto command executed by this function. This is mainly intended for
advanced usage and useful for CLI arguments which are not yet mirrored in a
dedicated parameter of this \R function. See \verb{quarto render --help} for options.}
}
\description{
Remove an extension in this folder or project by running \verb{quarto remove}
}
\section{Extension Trust}{
Quarto extensions may execute code when documents are rendered. Therefore, if
you do not trust the author of an extension, we recommend that you do not
install or use the extension.
By default \code{no_prompt = FALSE} which means that
the function will ask for explicit approval when used interactively, or
disallow installation.
}

\examples{
\dontrun{
# Remove an already installed extension
quarto_remove_extension("quarto-ext/fontawesome")
}
}
