% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{quarto_version}
\alias{quarto_version}
\title{Check quarto version}
\usage{
quarto_version()
}
\value{
a \code{\link[base:numeric_version]{numeric_version}} with the quarto version found
}
\description{
Determine the specific version of quartobinary found by \code{\link[=quarto_path]{quarto_path()}}.
If it returns \verb{99.9.9} then it means you are using a dev version.
}
