% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{quarto_list_extensions}
\alias{quarto_list_extensions}
\title{List Installed Quarto extensions}
\usage{
quarto_list_extensions()
}
\value{
A data frame with the installed extensions or NULL (invisibly) if no extensions are installed.
}
\description{
List Quarto Extensions in this folder or project by running \verb{quarto list}
}
\examples{
\dontrun{
# List Quarto Extensions in this folder or project
quarto_list_extensions()
}

}
