% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{quarto_path}
\alias{quarto_path}
\title{Path to the quarto binary}
\usage{
quarto_path()
}
\value{
Path to quarto binary (or \code{NULL} if not found)
}
\description{
Determine the path to the quarto binary. Uses \code{QUARTO_PATH} environment
variable if defined, otherwise uses \code{Sys.which()}.
}
