% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{quarto_use_template}
\alias{quarto_use_template}
\title{Use acustom format extension template}
\usage{
quarto_use_template(template, no_prompt = FALSE)
}
\arguments{
\item{template}{The template to install, either an archive or a GitHub
repository as described in the documentation
\url{https://quarto.org/docs/extensions/formats.html}.}

\item{no_prompt}{Do not prompt to confirm approval to download external extension.}
}
\description{
Install and use a template for Quarto using \verb{quarto use}.
}
\examples{
\dontrun{
# Install a template and set up a draft document from a GitHub repository
quarto_use_template("quarto-journals/jss")

# Install a template and set up a draft document from a ZIP archive
quarto_use_template("https://github.com/quarto-journals/jss/archive/refs/heads/main.zip")
}

}
