% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{quarto_use_template}
\alias{quarto_use_template}
\title{Quarto Use Template}
\usage{
quarto_use_template(template = NULL)
}
\arguments{
\item{template}{The template to install, either an archive or a GitHub
repository as described in the documentation
\url{https://quarto.org/docs/extensions/formats.html}.}
}
\description{
Install and use a template into a Quarto project.
}
\examples{
\dontrun{
# Install a template and set up a draft document from a GitHub repository
quarto_use_template("quarto-journals/jss")

# Install a template and set up a draft document from a ZIP archive
quarto_use_template("https://github.com/quarto-journals/jss/archive/refs/heads/main.zip")
}

}
